test: ♻️ Refacto modal and drawer components for e2e tests
