+++ apps/ledger-live-desktop/tests/page/drawer/swap.confirmation.drawer.ts
@@ -1,6 +1,6 @@
 import { expect } from "@playwright/test";
 import { step } from "tests/misc/reporters/step";
-import { Drawer } from "tests/page/drawer/drawer";
+import { Drawer } from "tests/component/drawer.component";
 import { capitalizeFirstLetter } from "tests/utils/textParserUtils";
 
 export class SwapConfirmationDrawer extends Drawer {
