+++ apps/ledger-live-desktop/tests/specs/services/confirmTransaction.spec.ts
@@ -3,7 +3,7 @@ import { expect } from "@playwright/test";
 import { Modal } from "../../component/modal.component";
 import { DiscoverPage } from "../../page/discover.page";
 import { Layout } from "../../component/layout.component";
-import { Drawer } from "../../page/drawer/drawer";
+import { Drawer } from "../../component/drawer.component";
 import { DeviceAction } from "../../models/DeviceAction";
 import { randomUUID } from "crypto";
 import { LiveAppWebview } from "../../models/LiveAppWebview";
