+++ apps/ledger-live-desktop/tests/specs/services/cosmosStaking.spec.ts
@@ -1,6 +1,6 @@
 import test from "../../fixtures/common";
 import { expect } from "@playwright/test";
-import { Drawer } from "../../page/drawer/drawer";
+import { Drawer } from "../../component/drawer.component";
 import { Modal } from "../../component/modal.component";
 import { PortfolioPage } from "../../page/portfolio.page";
 import { Layout } from "../../component/layout.component";
